backend/requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.16
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
idna==2.10
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.1
Pygments==2.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.11
wrapt==1.12.1
